AutoTest_Backend/pyproject.toml
[project]
name = "autotest-backend"
version = "0.1.0"
description = "Backend for the AutoTest Agent graduation project."
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.128.0",
    "uvicorn==0.40.0",
    "openai==2.15.0",
    "chromadb==1.4.1",
    "pydantic-settings==2.12.0",
    "selenium==4.40.0",
]

[tool.uv]
package = false
